inconvenient_resume_api/pyproject.toml
[tool.poetry]
name = "inconvenient-resume-api"
version = "0.1.0"
description = ""
authors = ["gage russell"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
pydantic = "^2.8.2"
sqlmodel = "^0.0.21"
pydantic-settings = "^2.4.0"
passlib = "^1.7.4"
email-validator = "^2.2.0"
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
uvicorn = "^0.30.5"
bcrypt = "^4.2.0"
alembic = "^1.13.2"
psycopg-binary = "^3.2.1"
psycopg = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
